TFS-274: Tilpassninger iht nye abac policies.
