+++ lokalutvikling/vtp.env
@@ -1,1 +1,2 @@
 CREATE_TOPICS=privat-familie-vedtakFattet-v1-vtp
+no.nav.modig.security.appcert.password=changeit
