+++ web/src/main/java/no/nav/foreldrepenger/abakus/app/sikkerhet/PdpRequestBuilderImpl.java
@@ -20,7 +20,6 @@ import no.nav.vedtak.sikkerhet.abac.StandardAbacAttributtType;
 public class PdpRequestBuilderImpl implements PdpRequestBuilder {
 
     public static final String ABAC_DOMAIN = "duplo";
-    public static final String PEP_ID = "abakus";
 
     @Override
     public PdpRequest lagPdpRequest(AbacAttributtSamling attributter) {
@@ -28,7 +27,6 @@ public class PdpRequestBuilderImpl implements PdpRequestBuilder {
         pdpRequest.put(PdpKlient.ENVIRONMENT_AUTH_TOKEN, attributter.getIdToken());
         pdpRequest.put(AbacAttributter.XACML_1_0_ACTION_ACTION_ID, attributter.getActionType().getEksternKode());
         pdpRequest.put(AbacAttributter.RESOURCE_FELLES_DOMENE, ABAC_DOMAIN);
-        pdpRequest.put(AbacAttributter.RESOURCE_FELLES_PEP_ID, PEP_ID);
         pdpRequest.put(AbacAttributter.RESOURCE_FELLES_RESOURCE_TYPE, attributter.getResource());
         pdpRequest.put(AbacAttributter.RESOURCE_FELLES_PERSON_AKTOERID_RESOURCE, attributter.getVerdier(StandardAbacAttributtType.AKTØR_ID));
         pdpRequest.put(AbacAttributter.RESOURCE_FELLES_PERSON_FNR, attributter.getVerdier(StandardAbacAttributtType.FNR));
